// src/api/type/Mrc721ItemBean.ts
export interface Mrc721ItemBean {
  chain: string;
  collectionPinId: string;
  collectionName: string;
  itemPinId: string;
  itemPinNumber: number;
  descPinId: string;
  name: string;
  desc: string;
  cover: string;
  metaData: string;
  createTime: number;
  address: string;
  content: string;
  metaId: string;
  descadded: boolean;
  contentType: string;
  contentTypeDetect: string;
  contentString: string;
  outpoint: string;
  outValue: number;
  showImage:string;
}